refactor: extract withAuthAction wrapper + shared createLocalStorageMock
Follow-up dedupe across server actions + test helpers:

- New `src/lib/auth-wrapper.ts` with `withAuthAction<T>(fn)`: funnels
  all six listening-queue / player-session actions through the same
  auth check so the `{ success: false, error: "Unauthorized" }` guard
  can't be silently forgotten in a new action. The wrapper is a plain
  helper (no "use server" directive), so importing it into the action
  modules keeps it out of the server-action surface.
- New `src/test/mocks/local-storage.ts` with `createLocalStorageMock()`:
  removes four verbatim copies across `migration-marker.test.ts`,
  `queue-persistence.test.ts`, `player-session.test.ts`, and
  `use-pwa-install.test.ts`. The four copies were already drifting
  (semicolon / punctuation nits) — single source now.

Net: -27 LOC across test files, +58 LOC across helpers, action bodies
shrink by one level of auth-guard indentation. All 1821 tests pass,
lint + build clean.
