+++ src/app/actions/listening-queue.ts
@@ -1,10 +1,10 @@
 "use server"
 
-import { auth } from "@clerk/nextjs/server"
 import { eq, asc } from "drizzle-orm"
 import { db } from "@/db"
 import { ensureUserExists } from "@/db/helpers"
 import { userQueueItems } from "@/db/schema"
+import { withAuthAction } from "@/lib/auth-wrapper"
 import {
   queueSchema,
   toAudioEpisode,
@@ -15,22 +15,19 @@ import {
 export async function getQueue(): Promise<
   { success: true; data: AudioEpisode[] } | { success: false; error: string }
 > {
-  const { userId } = await auth()
-  if (!userId) return { success: false, error: "Unauthorized" }
-
-  try {
-    const rows = await db.query.userQueueItems.findMany({
-      where: eq(userQueueItems.userId, userId),
-      orderBy: [asc(userQueueItems.position)],
-    })
-
-    const data: AudioEpisode[] = rows.map(toAudioEpisode)
-
-    return { success: true, data }
-  } catch (e) {
-    console.error("Failed to get queue:", e)
-    return { success: false, error: "Failed to get queue" }
-  }
+  return withAuthAction(async (userId) => {
+    try {
+      const rows = await db.query.userQueueItems.findMany({
+        where: eq(userQueueItems.userId, userId),
+        orderBy: [asc(userQueueItems.position)],
+      })
+      const data: AudioEpisode[] = rows.map(toAudioEpisode)
+      return { success: true as const, data }
+    } catch (e) {
+      console.error("Failed to get queue:", e)
+      return { success: false as const, error: "Failed to get queue" }
+    }
+  })
 }
 
 /**
@@ -47,50 +44,48 @@ export async function getQueue(): Promise<
 export async function setQueue(
   episodes: AudioEpisode[]
 ): Promise<{ success: true } | { success: false; error: string }> {
-  const { userId } = await auth()
-  if (!userId) return { success: false, error: "Unauthorized" }
+  return withAuthAction(async (userId) => {
+    const parsed = queueSchema.safeParse(episodes)
+    if (!parsed.success) {
+      console.warn("[setQueue] validation failed", parsed.error.issues)
+      return { success: false as const, error: "Invalid queue data" }
+    }
 
-  const parsed = queueSchema.safeParse(episodes)
-  if (!parsed.success) {
-    console.warn("[setQueue] validation failed", parsed.error.issues)
-    return { success: false, error: "Invalid queue data" }
-  }
+    try {
+      await ensureUserExists(userId)
 
-  try {
-    await ensureUserExists(userId)
+      await db.transaction(async (tx) => {
+        await tx.delete(userQueueItems).where(eq(userQueueItems.userId, userId))
 
-    await db.transaction(async (tx) => {
-      await tx.delete(userQueueItems).where(eq(userQueueItems.userId, userId))
+        if (parsed.data.length > 0) {
+          const rows = parsed.data.map((ep, index) => ({
+            userId,
+            position: index,
+            ...toEpisodeDenormRow(ep),
+            updatedAt: new Date(),
+          }))
+          await tx.insert(userQueueItems).values(rows)
+        }
+      })
 
-      if (parsed.data.length > 0) {
-        const rows = parsed.data.map((ep, index) => ({
-          userId,
-          position: index,
-          ...toEpisodeDenormRow(ep),
-          updatedAt: new Date(),
-        }))
-        await tx.insert(userQueueItems).values(rows)
-      }
-    })
-
-    return { success: true }
-  } catch (e) {
-    console.error("Failed to set queue:", e)
-    return { success: false, error: "Failed to set queue" }
-  }
+      return { success: true as const }
+    } catch (e) {
+      console.error("Failed to set queue:", e)
+      return { success: false as const, error: "Failed to set queue" }
+    }
+  })
 }
 
 export async function clearQueue(): Promise<
   { success: true } | { success: false; error: string }
 > {
-  const { userId } = await auth()
-  if (!userId) return { success: false, error: "Unauthorized" }
-
-  try {
-    await db.delete(userQueueItems).where(eq(userQueueItems.userId, userId))
-    return { success: true }
-  } catch (e) {
-    console.error("Failed to clear queue:", e)
-    return { success: false, error: "Failed to clear queue" }
-  }
+  return withAuthAction(async (userId) => {
+    try {
+      await db.delete(userQueueItems).where(eq(userQueueItems.userId, userId))
+      return { success: true as const }
+    } catch (e) {
+      console.error("Failed to clear queue:", e)
+      return { success: false as const, error: "Failed to clear queue" }
+    }
+  })
 }
