+++ src/hooks/__tests__/use-pwa-install.test.ts
@@ -18,6 +18,7 @@ vi.mock("next/navigation", () => ({
 }));
 
 import { usePwaInstall } from "@/hooks/use-pwa-install";
+import { createLocalStorageMock } from "@/test/mocks/local-storage";
 
 function createMockMatchMedia(overrides: Record<string, boolean> = {}) {
   return (query: string): MediaQueryList => ({
@@ -43,27 +44,6 @@ function fireBeforeInstallPrompt(): BeforeInstallPromptEvent {
   Object.defineProperty(event, "platforms", { value: ["web"] });
   window.dispatchEvent(event);
   return event;
-}
-
-// Simple in-memory localStorage mock that has all standard methods
-function createLocalStorageMock(): Storage {
-  const store: Record<string, string> = {};
-  return {
-    getItem: (key: string) => store[key] ?? null,
-    setItem: (key: string, value: string) => {
-      store[key] = value;
-    },
-    removeItem: (key: string) => {
-      delete store[key];
-    },
-    clear: () => {
-      for (const key of Object.keys(store)) delete store[key];
-    },
-    get length() {
-      return Object.keys(store).length;
-    },
-    key: (index: number) => Object.keys(store)[index] ?? null,
-  };
 }
 
 describe("usePwaInstall", () => {
