+++ src/lib/__tests__/migration-marker.test.ts
@@ -8,26 +8,7 @@ import {
   markSessionMigrated,
   setLastUserId,
 } from "@/lib/migration-marker"
-
-function createLocalStorageMock(): Storage {
-  const store: Record<string, string> = {}
-  return {
-    getItem: (key: string) => store[key] ?? null,
-    setItem: (key: string, value: string) => {
-      store[key] = value
-    },
-    removeItem: (key: string) => {
-      delete store[key]
-    },
-    clear: () => {
-      for (const key of Object.keys(store)) delete store[key]
-    },
-    get length() {
-      return Object.keys(store).length
-    },
-    key: (index: number) => Object.keys(store)[index] ?? null,
-  }
-}
+import { createLocalStorageMock } from "@/test/mocks/local-storage"
 
 describe("migration-marker", () => {
   beforeEach(() => {
