+++ src/lib/__tests__/queue-persistence.test.ts
@@ -1,6 +1,7 @@
 import { describe, it, expect, beforeEach } from "vitest"
 import { loadQueue, saveQueue } from "@/lib/queue-persistence"
 import type { AudioEpisode } from "@/contexts/audio-player-context"
+import { createLocalStorageMock } from "@/test/mocks/local-storage"
 
 const validEpisode: AudioEpisode = {
   id: "ep-1",
@@ -16,26 +17,6 @@ const validEpisode2: AudioEpisode = {
   title: "Second Episode",
   podcastTitle: "Another Podcast",
   audioUrl: "https://example.com/audio2.mp3",
-}
-
-function createLocalStorageMock(): Storage {
-  const store: Record<string, string> = {}
-  return {
-    getItem: (key: string) => store[key] ?? null,
-    setItem: (key: string, value: string) => {
-      store[key] = value
-    },
-    removeItem: (key: string) => {
-      delete store[key]
-    },
-    clear: () => {
-      for (const key of Object.keys(store)) delete store[key]
-    },
-    get length() {
-      return Object.keys(store).length
-    },
-    key: (index: number) => Object.keys(store)[index] ?? null,
-  }
 }
 
 describe("loadQueue", () => {
